Created TempEnvironment.java and LIQUIDEnvironment
I've ported over the testbed LIQUID class and its superclass
TestbedTest.  I have not yet been able to test that these work properly,
and I had to remove all of the serialization code, as well as any
references to the testbed settings.  I have arbitrarily filled in the
testbed settings with values