KNOWN BUG
Trying to load from a log file and run it without making changes
generates a null pointer exception in the file writer class, as bw is
not initialized and wipes the log file.  I have tried a few things to
remedy this, but none have worked.  I am unsure how to do this, but
re-writing the parameters to the log file should work as a temporary
fix.